+++ projects/net.splitcells.gel.ext/CHANGELOG.md
@@ -5,6 +5,8 @@
 # Changelog
 ## [Unreleased]
 ### Major Changes
+* **2025-11-19 \#37**: Rename `HasMinimalDistanceOfCallRunner#has_minimal_distance_of` to `HasMinimalDistanceOfCallRunner#hasMinimalDistanceOf`,
+  in order to standardize function naming.
 * **2022-10-16 \#170**: Convert `DatabaseIRefFactory` to database aspect and rename it to `DatabaseMetaAspect`.
 ### Minor Changes
 * **2023-02-26 \#170**: The development config `GelDev` now starts the web server in the background,